+++ packages/http-client-java/generator/http-client-generator-test/src/test/java/azure/clientgenerator/core/usage/UsageTests.java
@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package _specs_.azure.clientgenerator.core.usage;
+package azure.clientgenerator.core.usage;
 
-import _specs_.azure.clientgenerator.core.usage.modelinoperation.models.InputModel;
-import _specs_.azure.clientgenerator.core.usage.modelinoperation.models.OutputModel;
-import _specs_.azure.clientgenerator.core.usage.modelinoperation.models.ResultModel;
-import _specs_.azure.clientgenerator.core.usage.modelinoperation.models.RoundTripModel;
+import azure.clientgenerator.core.usage.models.InputModel;
+import azure.clientgenerator.core.usage.models.OutputModel;
+import azure.clientgenerator.core.usage.models.ResultModel;
+import azure.clientgenerator.core.usage.models.RoundTripModel;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Modifier;
 import org.junit.jupiter.api.Assertions;
