+++ packages/http-client-java/generator/http-client-generator-test/src/test/java/azure/core/scalar/ScalarTests.java
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package _specs_.azure.core.scalar;
+package azure.core.scalar;
 
-import _specs_.azure.core.scalar.models.AzureLocationModel;
+import azure.core.scalar.models.AzureLocationModel;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
